+++ src/Exceptions/InvalidCharactersException.php
@@ -3,8 +3,10 @@
 namespace PragmaRX\Google2FA\Exceptions;
 
 use Exception;
+use PragmaRX\Google2FA\Exceptions\Contracts\Google2FA as Google2FAExceptionContract;
+use PragmaRX\Google2FA\Exceptions\Contracts\InvalidCharacters as InvalidCharactersExceptionContract;
 
-class InvalidCharactersException extends Exception
+class InvalidCharactersException extends Google2FAException implements Google2FAExceptionContract, InvalidCharactersExceptionContract
 {
     protected $message = 'Invalid characters in the base32 string.';
 }