+++ changelog.md
@@ -1,5 +1,13 @@
 ## Change Log
 
+## [5.1.0] - 2019-06-23
+### Added
+- Base exception class and interfaces 
+
+## [5.0.0] - 2019-05-19
+### Changed
+- Remove dead Google Charts API 
+
 ## [4.0.0] - 2018-10-06
 ### Changed
 - Bacon QRCode package removed